requirements.txt
fastapi
uvicorn
requests
google-generativeai
sqlalchemy
python-dotenv
bandit
pylint
psycopg2-binary